Data forms and tables
Added full functionality to add and remove elements from the server's people list. The server will automatically append the first and last name together to create a full name string as part of the object. There is also a working table of this data, which will automatically update each time the page is refreshed or the data is altered.